+++ rmq_metadata.py
@@ -780,6 +780,7 @@ def get_textract_data(f_name, cfg, log, **kwargs):
             log.log_info("get_textract_data:  New encoding code detected: %s" %
                          (char_encoding))
             rawtext = extract_pdf(f_name, char_encoding)
+            log.log_info("get_textract_data:  Re-running word_tokenizer...")
             tokens = word_tokenize(rawtext)
 
         else:
@@ -806,6 +807,7 @@ def pdf_to_string(f_name, log):
     Arguments:
         (input) f_name -> PDF file name.
         (input) log -> Log class instance.
+        (output) status -> True|False - successfully extraction of data.
         (output) text -> Raw text.
 
     """
@@ -834,6 +836,41 @@ def pdf_to_string(f_name, log):
     text = data.replace('.','')
 
     return status, text
+
+
+def get_pdfminer_data(f_name, cfg, log):
+
+    """Function:  get_pdfminer_data
+
+    Description:  Process data using the pdfminer module.
+
+    Arguments:
+        (input) f_name -> PDF file name.
+        (input) cfg -> Configuration settings module for the program.
+        (input) log -> Log class instance.
+        (output) status -> True|False - successfully extraction of data.
+        (output) final_data -> List of categorized tokens from PDF file.
+
+    """
+
+    final_data = []
+    log.log_info("get_pdfminer_data:  Extracting data using pdfminer.")
+    status, rawtext = pdf_to_string(f_name, log)
+
+    if status:
+        log.log_info("get_pdfminer_data:  Running word_tokenizer...")
+        tokens = word_tokenize(rawtext)
+        log.log_info("get_pdfminer_data:  Finding tokens.")
+        categorized_text = find_tokens(tokens, cfg)
+
+        if categorized_text:
+            log.log_info("get_pdfminer_data:  Summarizing data")
+            final_data = summarize_data(categorized_text, cfg.token_types)
+
+    else:
+        final_data = []
+
+    return status, final_data
 
 
 def _process_queue(queue, body, r_key, cfg, f_name, log, **kwargs):